requirements.txt
streamlit
biopython
requests
openai
ete3
matplotlib
python-dotenv